+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
   - Fixed issue on Elasticsearch while logging java *Error classes. #98
   - Fixed issue on CloudWatch Logs when publishing empty payloads. #125
-  - Added integration tests with testcontainers for ELS, CloudWatch
+  - Added integration tests with testcontainers for ELS, CloudWatch, Kinesis
 
 ## v0.9.0 - (2022-07-11)